Compiler warnings removed.

+++ MasterSimulatorUI/src/MSIMViewSlaves.cpp
@@ -393,14 +393,14 @@ void MSIMViewSlaves::on_toolButtonAddSlave_clicked() {
 	// create a block for the graphical representation
 
 #if (QT_VERSION >= QT_VERSION_CHECK(5, 10, 0))
-	unsigned int gx = BLOCKMOD::Globals::GridSpacing*QRandomGenerator::global()->generateDouble()*40.0;
-	unsigned int gy = BLOCKMOD::Globals::GridSpacing*QRandomGenerator::global()->generateDouble()*40.0;
+	unsigned int gx = static_cast<unsigned int>(BLOCKMOD::Globals::GridSpacing*QRandomGenerator::global()->generateDouble()*40.0);
+	unsigned int gy = static_cast<unsigned int>(BLOCKMOD::Globals::GridSpacing*QRandomGenerator::global()->generateDouble()*40.0);
 #else
-	unsigned int gx = BLOCKMOD::Globals::GridSpacing*qrand()*40.0/RAND_MAX;
-	unsigned int gy = BLOCKMOD::Globals::GridSpacing*qrand()*40.0/RAND_MAX;
+	unsigned int gx = static_cast<unsigned int>(BLOCKMOD::Globals::GridSpacing*qrand()*40.0/RAND_MAX);
+	unsigned int gy = static_cast<unsigned int>(BLOCKMOD::Globals::GridSpacing*qrand()*40.0/RAND_MAX);
 #endif
 	BLOCKMOD::Block b(QString::fromStdString(simDef.m_name), gx, gy);
-	b.m_size = QSize(BLOCKMOD::Globals::GridSpacing*8, BLOCKMOD::Globals::GridSpacing*12);
+	b.m_size = QSize(static_cast<int>(BLOCKMOD::Globals::GridSpacing*8), static_cast<int>(BLOCKMOD::Globals::GridSpacing*12));
 	n.m_blocks.push_back(b);
 
 	// create undo action - this will update the network and also
